requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
dj-database-url==3.1.0
Django==6.0.1
django-import-export==4.4.0
gunicorn==25.0.1
numpy==2.4.2
packaging==26.0
pgvector==0.4.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tablib==3.9.0
whitenoise==6.11.0
djangorestframework
openai
djangorestframework-simplejwt